feat: CRUD para county, course, province

+++ src/use-cases/course/destroy-course.ts
@@ -1,4 +1,5 @@
 import { CoursesRepository } from '@/repositories/course-repository'
+import { CourseNotFoundError } from '../errors/course-not-found'
 
 interface DestroyCourseUseCaseRequest {
   id: number
@@ -10,6 +11,11 @@ export class DestroyCourseUseCase {
   async execute({
     id,
   }: DestroyCourseUseCaseRequest): Promise<Boolean> {
+    const findCourse = await this.coursesRepository.findById(id)
+
+    if(!findCourse){
+      throw new CourseNotFoundError()
+    }
     return await this.coursesRepository.destroy(
       id,
     )
